requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
cachetools==5.2.0
caveclient==4.17.1
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
idna==3.4
importlib-resources==5.10.2
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
neuvueclient==0.11.3
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
slack-bolt==1.16.1
slack-sdk==3.19.5
stack-data==0.6.2
tabulate==0.9.0
tqdm==4.64.1
traitlets==5.8.0
typing==3.7.4.3
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
zipp==3.11.0
